189 implemented ToolTipPause how to interact with flashlight
