+++ Assets/Scripts/InteractableObjects/Interfaces.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 631d2a0d12994d52857f2d4612436c28
+timeCreated: 1517068030
