+++ Assets/Scripts/InteractableObjects/Interfaces/IHowToolTipWithObject.cs
@@ -1,0 +1,7 @@
+﻿namespace InteractableObjects.Interfaces
+{
+    public interface IHowToolTipWithObject
+    {
+        void HotToInteractToolTip();
+    }
+}
