+++ Assets/Scripts/InteractableObjects/Interfaces/IHowToolTipWithObject.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 48f348fb5add4c7a88e693f11f960d53
+timeCreated: 1517067944
