+++ Assets/Scripts/InteractableObjects/Pickable.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 367bfed291a7446d93c5b37a5f88fe38
+timeCreated: 1517067248
