+++ Assets/Scripts/InteractableObjects/Pickable/FlashLight.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b4f173a09eea49aea390ef4859be9116
+timeCreated: 1513643573
